consolidate details pages
