+++ src/services/ui/src/pages/dashboard/Lists/waivers/consts.tsx
@@ -16,9 +16,7 @@ export const TABLE_COLUMNS = (props?: { isCms?: boolean }): OsTableColumn[] => [
       return (
         <Link
           className="cursor-pointer text-blue-600"
-          to={`/detail/${data.authority?.toLowerCase()}?id=${encodeURIComponent(
-            data.id
-          )}`}
+          to={`/details?id=${encodeURIComponent(data.id)}`}
         >
           {data.id}
         </Link>
