--- src/services/ui/src/pages/detail/index.ts
@@ -1,3 +1,0 @@
-export * from "./chip-spa";
-export * from "./medicaid-spa";
-export * from "./waiver";
