--- src/services/ui/src/pages/detail/medicaid-spa.tsx
@@ -1,9 +1,0 @@
-import { DetailWrapper } from "./wrapper";
-import { ChipSpa } from "./chip-spa";
-
-export const MedicaidSpaPage = () => (
-  <DetailWrapper>
-    {/* <MedicaidSpa /> */}
-    <ChipSpa />
-  </DetailWrapper>
-);
