--- src/services/ui/src/pages/detail/waiver.tsx
@@ -1,9 +1,0 @@
-import { DetailWrapper } from "./wrapper";
-import { ChipSpa } from "./chip-spa";
-
-export const WaiverPage = () => (
-  <DetailWrapper>
-    {/* <Waiver /> */}
-    <ChipSpa />
-  </DetailWrapper>
-);
