+++ src/services/ui/src/router.tsx
@@ -19,9 +19,7 @@ export const router = createBrowserRouter([
         element: <P.Dashboard />,
         loader: dashboardLoader(queryClient),
       },
-      { path: "/detail/chip-spa", element: <P.ChipSpaPage /> },
-      { path: "/detail/medicaid-spa", element: <P.MedicaidSpaPage /> },
-      { path: "/detail/waiver", element: <P.WaiverPage /> },
+      { path: "/details", element: <P.Details /> },
       { path: "/faq", element: <P.Faq /> },
     ],
     loader: rootLoader(queryClient),
